combine AttributeColorConverter and AttributeIconConverter code
